pyproject.toml
[project]
name = "clgo-curator"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "mlx-lm>=0.31.1",
    "mlx-tune>=0.4.15",
]
